revision tokencounter

+++ counter/state.go
@@ -1,7 +1,7 @@
 package counter
 
 type State struct {
-	PromptTokens     int
-	CompletionTokens int
-	TotalTokens      int
+	PromptTokens     int64
+	CompletionTokens int64
+	TotalTokens      int64
 }
